+++ src/api/config/privateApi.server.ts
@@ -1,36 +1,9 @@
-import axios, { InternalAxiosRequestConfig } from 'axios';
-import { cookies } from 'next/headers';
-
-import { AUTH_COOKIE_KEYS } from '~/types/auth';
-import { getAccessTokenServer } from '~/utils/auth/tokenValidator.server';
+import axios from 'axios';
 
 import { CustomInstance } from './api.types';
 import { onResponse } from './interceptor';
+import { onRequestServer } from './interceptor.server';
 import { ROOT_API_URL } from './requestUrl';
-
-const onServerRequest = async (config: InternalAxiosRequestConfig) => {
-  try {
-    const cookieStore = cookies();
-    const accessToken = cookieStore.get(AUTH_COOKIE_KEYS.accessToken)?.value;
-    const accessTokenExpireDate = Number(
-      cookieStore.get(AUTH_COOKIE_KEYS.accessTokenExpireDate)?.value,
-    );
-    const refreshToken = cookieStore.get(AUTH_COOKIE_KEYS.refreshToken)?.value;
-    const validToken = await getAccessTokenServer({
-      accessToken,
-      accessTokenExpireDate,
-      refreshToken,
-    });
-    if (validToken) {
-      config.headers.Authorization = `Bearer ${validToken.accessToken}`;
-      return config;
-    }
-    throw new Error('로그인이 필요합니다.');
-  } catch (error) {
-    // server-side 로그아웃 처리
-    return Promise.reject(error);
-  }
-};
 
 export const privateApi: CustomInstance = axios.create({
   baseURL: ROOT_API_URL,
@@ -38,6 +11,6 @@ export const privateApi: CustomInstance = axios.create({
 
 privateApi.defaults.timeout = 2500;
 
-privateApi.interceptors.request.use(onServerRequest);
+privateApi.interceptors.request.use(onRequestServer);
 
 privateApi.interceptors.response.use(onResponse);
