+++ src/api/config/privateApi.ts
@@ -3,7 +3,8 @@ import axios from 'axios';
 
 import { CustomInstance } from '~/api/config/api.types';
 
-import { onRequest, onRequestError, onResponse, onResponseError } from './interceptor';
+import { onRequestError, onResponse } from './interceptor';
+import { onRequestClient, onResponseErrorClient } from './interceptor.client';
 import { ROOT_API_URL } from './requestUrl';
 
 const privateApi: CustomInstance = axios.create({
@@ -12,8 +13,8 @@ const privateApi: CustomInstance = axios.create({
 
 privateApi.defaults.timeout = 2500;
 
-privateApi.interceptors.request.use(onRequest, onRequestError);
+privateApi.interceptors.request.use(onRequestClient, onRequestError);
 
-privateApi.interceptors.response.use(onResponse, onResponseError);
+privateApi.interceptors.response.use(onResponse, onResponseErrorClient);
 
 export default privateApi;
